+++ RELEASE-NOTES.md
@@ -11,6 +11,7 @@
 * The `Property` constructor now accepts an optional `StatementList` parameter
 * Added `Property::getStatements` and `Property::setStatements`
 * Added `StatementList::getAllSnaks` to use instead of `Entity::getAllSnaks`
+* Added `Statement::setClaim`
 
 ## Version 1.0 (2014-09-02)